US201764; support deploy office.zip and office-loader.zip into the web-dossier, add smitherfile to run test

+++ workspace-settings/rake/lib/tasks/test/integration.rb
@@ -6,8 +6,18 @@ end
 
 
 desc "package test docker"
-task :deploy_tester_server do
+task :deploy_tester_server,[:build_no] do | t, args|
+  
+  group_id = Common::Version.dependency_group_id
+  artifact_id = "office-loader"
+  version = args['build_no'] || Nexus.latest_artifact_version(artifact_id: artifact_id, group_id: group_id)
+  download_mstr_office(group_id, version)
+  download_latest_web_dossier
+  
+  do_stop_local_web_dossier
   do_package_test_docker
+  do_start_local_web_dossier
+
   # cmd = "docker build -t mstr-dossier ."
   # shell_command! cmd, cwd: "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/e2e-webserver"
   
@@ -17,13 +27,15 @@ desc "package test docker"
 task :e2e_test_browser do
  
   shell_command! "npm install", cwd: "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/test-driver-browser"
-  shell_command! "npm run test", cwd: "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/test-driver-borwser"
+  shell_command! "npm run test", cwd: "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/test-driver-browser"
 
 end
 
 desc "package test docker"
 task :tp do
   stop_test_web_service
+  download_latest_web_dossier
+  download_mstr_office
   do_package_test_docker
   # cmd = "docker build -t mstr-dossier ."
   # shell_command! cmd, cwd: "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/e2e-webserver"
@@ -31,7 +43,6 @@ task :tp do
 end
 
 def do_package_test_docker()
-  download_latest_web_dossier
   cmd = "docker build -t mstr-dossier ."
   shell_command! cmd, cwd: "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/e2e-webserver"
 end
@@ -44,26 +55,52 @@ def stop_test_web_service()
 end
 def download_latest_web_dossier()
   dossior_war_path = "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/e2e-webserver/web-dossier.war"
-  if File.exist? dossior_war_path
-    FileUtils.rm_rf dossior_war_path
-  end
-  Nexus.download_latest_artifact(file_path:dossior_war_path,artifact_id:'web-dossier',group_id:Common::Version.dependency_group_id,repository:'releases',extra_coordinates:{e: 'war'})
+  FileUtils.rm_rf dossior_war_path if File.exist? dossior_war_path
+  Nexus.download_latest_artifact(file_path:dossior_war_path,
+    artifact_id:'web-dossier',
+    group_id:Common::Version.dependency_group_id,
+    repository:'releases',
+    extra_coordinates:{e: 'war'})
 
+end
+
+def download_mstr_office(group_id, version)
+  artifacts = ["office-loader", "office"]
+  artifacts.each do |artifact_id|
+    mstr_office_path = "#{$WORKSPACE_SETTINGS[:paths][:project][:tests][:home]}/integration/e2e-webserver/#{artifact_id}.zip"
+    FileUtils.rm_rf mstr_office_path if File.exist? mstr_office_path
+    Nexus.download_artifact(file_path:mstr_office_path,
+      artifact_id:artifact_id,
+      version: version,
+      group_id:group_id,
+      repository:'releases',
+      extra_coordinates:{e: 'zip'})
+  end
+  
 end
 
 
 desc "package test docker"
 task :r do
   # cmd = "docker run -it -p 8888:8080 --name office-tester mstr-dossier /bin/bash"
+  do_start_local_web_dossier
+end
+
+def do_start_local_web_dossier
   cmd = "docker run --rm -p 8080:8080 -p 8443:8443 --name office-tester -d mstr-dossier"
   shell_command! cmd
   wait_web_dossier_online
   good "web server online"
-  puts "please open http://localhost:8080/web-dossier/auth/ui/loginPage"
+  puts "please open https://localhost:8443/web-dossier/auth/ui/loginPage"
 end
 
-desc "run docker"
+
+desc "termiate docker container"
 task :k do
+  do_stop_local_web_dossier
+end
+
+def do_stop_local_web_dossier()
   cmd = "docker stop office-tester"
   shell_command cmd
   cmd = "docker container rm office-tester"
